# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Multi-Brand Shopify E-Commerce & B2B Wholesale Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/products:
    get:
      operationId: listAdminProducts
      tags: [admin]
      summary: List all products for the store
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [admin]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /admin/products/{productId}:
    get:
      operationId: getAdminProduct
      tags: [admin]
      summary: Get a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [admin]
      summary: Update a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [admin]
      summary: Delete a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /admin/products/{productId}/sync-shopify:
    post:
      operationId: syncProductToShopify
      tags: [admin]
      summary: Sync product to Shopify
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShopifySyncResult"
...
  /admin/categories:
    get:
      operationId: listAdminCategories
      tags: [admin]
      summary: List all categories
      responses:
        "200":
          description: Categories list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [admin]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /admin/categories/{categoryId}:
    patch:
      operationId: updateCategory
      tags: [admin]
      summary: Update a category
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryUpdate"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: deleteCategory
      tags: [admin]
      summary: Delete a category
      parameters:
        - name: categoryId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /admin/b2b-accounts:
    get:
      operationId: listB2BClients
      tags: [admin]
      summary: List all B2B wholesale clients
      responses:
        "200":
          description: B2B clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/B2BClient"
    post:
      operationId: createB2BClient
      tags: [admin]
      summary: Provision a new B2B wholesale client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/B2BClientInput"
      responses:
        "201":
          description: B2B client created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/B2BClient"
  /admin/b2b-accounts/{clientId}:
    get:
      operationId: getB2BClient
      tags: [admin]
      summary: Get a B2B client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: B2B client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/B2BClient"
    patch:
      operationId: updateB2BClient
      tags: [admin]
      summary: Update a B2B client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/B2BClientUpdate"
      responses:
        "200":
          description: Updated B2B client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/B2BClient"
    delete:
      operationId: deleteB2BClient
      tags: [admin]
      summary: Delete a B2B client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
  /admin/b2b-accounts/{clientId}/products:
    get:
      operationId: getB2BClientProducts
      tags: [admin]
      summary: Get product IDs assigned to a B2B client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Assigned product IDs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientProductAccess"
    put:
      operationId: setB2BClientProducts
      tags: [admin]
      summary: Set product access for a B2B client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientProductAccess"
      responses:
        "200":
          description: Updated access
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientProductAccess"
...
  /admin/orders/{orderId}:
    get:
      operationId: getAdminOrder
      tags: [admin]
      summary: Get an order with full details
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /admin/orders/{orderId}/status:
    patch:
      operationId: updateOrderStatus
      tags: [admin]
      summary: Advance order fulfillment status (step 1-4)
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ─── B2B PORTAL ─────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, storeId, name, price, status, channel, preOrder, createdAt]
      properties:
        id:
          type: number
        storeId:
          type: number
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        compareAtPrice:
          type: ["number", "null"]
        status:
          type: string
          enum: [active, disabled]
        channel:
          type: string
          enum: [all, b2b, b2c]
        preOrder:
          type: boolean
        preOrderNotice:
          type: ["string", "null"]
        shopifyProductId:
          type: ["string", "null"]
        shopifySynced:
          type: boolean
        variants:
          type: array
          items:
            $ref: "#/components/schemas/ProductVariant"
        images:
          type: array
          items:
            $ref: "#/components/schemas/ProductImage"
        categories:
          type: array
          items:
            type: number
        createdAt:
          type: string
    ProductVariant:
      type: object
      required: [id, productId, sku, inventory]
      properties:
        id:
          type: number
        productId:
          type: number
        color:
          type: ["string", "null"]
        size:
          type: ["string", "null"]
        sku:
          type: string
        inventory:
          type: number
        price:
          type: ["number", "null"]
    ProductImage:
      type: object
      required: [id, productId, url, displayOrder]
      properties:
        id:
          type: number
        productId:
          type: number
        url:
          type: string
        altText:
          type: ["string", "null"]
        displayOrder:
          type: number
    ProductInput:
      type: object
      required: [name, price]
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        compareAtPrice:
          type: number
        status:
          type: string
          enum: [active, disabled]
        channel:
          type: string
          enum: [all, b2b, b2c]
        preOrder:
          type: boolean
        preOrderNotice:
          type: string
        variants:
          type: array
          items:
            $ref: "#/components/schemas/VariantInput"
        images:
          type: array
          items:
            $ref: "#/components/schemas/ImageInput"
        categoryIds:
          type: array
          items:
            type: number
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        compareAtPrice:
          type: number
        status:
          type: string
          enum: [active, disabled]
        channel:
          type: string
          enum: [all, b2b, b2c]
        preOrder:
          type: boolean
        preOrderNotice:
          type: string
        variants:
          type: array
          items:
            $ref: "#/components/schemas/VariantInput"
        images:
          type: array
          items:
            $ref: "#/components/schemas/ImageInput"
        categoryIds:
          type: array
          items:
            type: number
    VariantInput:
      type: object
      required: [sku, inventory]
      properties:
        id:
          type: number
        color:
          type: string
        size:
          type: string
        sku:
          type: string
        inventory:
          type: number
        price:
          type: number
    ImageInput:
      type: object
      required: [url]
      properties:
        id:
          type: number
        url:
          type: string
        altText:
          type: string
        displayOrder:
          type: number
    ShopifySyncResult:
      type: object
      required: [success, message]
      properties:
        success:
          type: boolean
        message:
          type: string
        synced:
          type: ["number", "null"]
        errors:
          type: ["number", "null"]

    # ── Category ──────────────────────────────────────────────
    Category:
      type: object
      required: [id, storeId, name, displayOrder, productCount]
      properties:
        id:
          type: number
        storeId:
          type: number
        name:
          type: string
        parentId:
          type: ["number", "null"]
        displayOrder:
          type: number
        bannerImageUrl:
          type: ["string", "null"]
        productCount:
          type: number
    CategoryInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        parentId:
          type: number
        displayOrder:
          type: number
        bannerImageUrl:
          type: string
    CategoryUpdate:
      type: object
      properties:
        name:
          type: string
        parentId:
          type: number
        displayOrder:
          type: number
        bannerImageUrl:
          type: string

    # ── B2B Client ────────────────────────────────────────────
    B2BClient:
      type: object
      required: [id, storeId, email, companyName, discountPercent, paymentTerms, forcePasswordChange, isActive, createdAt]
      properties:
        id:
          type: number
        storeId:
          type: number
        email:
          type: string
        companyName:
          type: string
        contactName:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        discountPercent:
          type: number
        paymentTerms:
          type: string
          enum: [cod, net30]
        forcePasswordChange:
          type: boolean
        isActive:
          type: boolean
        assignedProductIds:
          type: array
          items:
            type: number
        createdAt:
          type: string
    B2BClientInput:
      type: object
      required: [email, password, companyName, discountPercent, paymentTerms]
      properties:
        email:
          type: string
        password:
          type: string
        companyName:
          type: string
        contactName:
          type: string
        phone:
          type: string
        discountPercent:
          type: number
        paymentTerms:
          type: string
          enum: [cod, net30]
    B2BClientUpdate:
      type: object
      properties:
        companyName:
          type: string
        contactName:
          type: string
        phone:
          type: string
        discountPercent:
          type: number
        paymentTerms:
          type: string
          enum: [cod, net30]
        isActive:
          type: boolean
    ClientProductAccess:
      type: object
      required: [productIds]
      properties:
        productIds:
          type: array
          items:
            type: number
...
    Order:
      type: object
      required: [id, storeId, type, status, fulfillmentStep, total, paymentTerms, createdAt]
      properties:
        id:
          type: number
        storeId:
          type: number
        storeName:
          type: ["string", "null"]
        type:
          type: string
          enum: [b2b, b2c]
        status:
          type: string
          enum: [received, production, shipped, delivered]
        fulfillmentStep:
          type: number
        total:
          type: number
        paymentTerms:
          type: string
          enum: [cod, net30, card]
        b2bClientId:
          type: ["number", "null"]
        b2bCompanyName:
          type: ["string", "null"]
        customerName:
          type: ["string", "null"]
        customerEmail:
          type: ["string", "null"]
        shippingAddress:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        shopifyOrderId:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
...
    OrderStatusUpdate:
      type: object
      required: [status, fulfillmentStep]
      properties:
        status:
          type: string
          enum: [received, production, shipped, delivered]
        fulfillmentStep:
          type: number

    # ── Analytics ─────────────────────────────────────────────
...